+++ src/window/basic_frame.rs
@@ -4,12 +4,10 @@ use std::sync::{Arc, Mutex};
 
 use wayland_client::commons::Implementation;
 use wayland_client::protocol::{
-    wl_buffer, wl_compositor, wl_pointer, wl_seat, wl_shm, wl_subcompositor, wl_subsurface,
-    wl_surface,
+    wl_compositor, wl_pointer, wl_seat, wl_shm, wl_subcompositor, wl_subsurface, wl_surface,
 };
 use wayland_client::Proxy;
 
-use wayland_client::protocol::wl_buffer::RequestsTrait as BufferRequests;
 use wayland_client::protocol::wl_compositor::RequestsTrait as CompositorRequests;
 use wayland_client::protocol::wl_pointer::RequestsTrait as PointerRequests;
 use wayland_client::protocol::wl_seat::RequestsTrait as SeatRequests;
@@ -216,7 +214,6 @@ fn find_button(x: f64, y: f64, w: u32) -> Location {
 pub struct BasicFrame {
     inner: Arc<Inner>,
     pools: DoubleMemPool,
-    buffers: Vec<Proxy<wl_buffer::WlBuffer>>,
     active: bool,
     hidden: bool,
     pointers: Vec<AutoPointer>,
@@ -233,7 +230,6 @@ impl Frame for BasicFrame {
         shm: &Proxy<wl_shm::WlShm>,
         implementation: Box<Implementation<u32, FrameRequest> + Send>,
     ) -> Result<BasicFrame, ::std::io::Error> {
-        let pools = DoubleMemPool::new(&shm)?;
         let parts = [
             Part::new(base_surface, compositor, subcompositor),
             Part::new(base_surface, compositor, subcompositor),
@@ -241,16 +237,26 @@ impl Frame for BasicFrame {
             Part::new(base_surface, compositor, subcompositor),
             Part::new(base_surface, compositor, subcompositor),
         ];
+        let inner = Arc::new(Inner {
+            parts: parts,
+            size: Mutex::new((1, 1)),
+            resizable: Mutex::new(true),
+            implem: Mutex::new(implementation),
+            maximized: Mutex::new(false),
+        });
+        let my_inner = inner.clone();
+        // Send a Refresh request on callback from DoubleMemPool as it will be fired when
+        // None was previously returned from `pool()` and the draw was postponed
+        let pools = DoubleMemPool::new(&shm, move |_, _| {
+            my_inner
+                .implem
+                .lock()
+                .unwrap()
+                .receive(FrameRequest::Refresh, 0);
+        })?;
         Ok(BasicFrame {
-            inner: Arc::new(Inner {
-                parts: parts,
-                size: Mutex::new((1, 1)),
-                resizable: Mutex::new(true),
-                implem: Mutex::new(implementation),
-                maximized: Mutex::new(false),
-            }),
+            inner,
             pools,
-            buffers: Vec::new(),
             active: false,
             hidden: false,
             pointers: Vec::new(),
@@ -391,15 +397,13 @@ impl Frame for BasicFrame {
             return;
         }
         let (width, height) = *(self.inner.size.lock().unwrap());
-        // destroy current pending buffers
-        // TODO: do double-buffering
-        for b in self.buffers.drain(..) {
-            b.destroy();
-        }
 
         {
             // grab the current pool
-            let pool = self.pools.pool();
+            let pool = match self.pools.pool() {
+                Some(pool) => pool,
+                None => return,
+            };
             // resize the pool as appropriate
             let pxcount = (HEADER_SIZE * width)
                 + max(
@@ -472,14 +476,13 @@ impl Frame for BasicFrame {
 
             // Create the buffers
             // -> head-subsurface
-            let buffer =
-                pool.buffer(
-                    0,
-                    width as i32,
-                    HEADER_SIZE as i32,
-                    4 * width as i32,
-                    wl_shm::Format::Argb8888,
-                ).implement(|_, _| {});
+            let buffer = pool.buffer(
+                0,
+                width as i32,
+                HEADER_SIZE as i32,
+                4 * width as i32,
+                wl_shm::Format::Argb8888,
+            );
             self.inner.parts[HEAD]
                 .subsurface
                 .set_position(0, -(HEADER_SIZE as i32));
@@ -499,17 +502,15 @@ impl Frame for BasicFrame {
                     .damage(0, 0, width as i32, HEADER_SIZE as i32);
             }
             self.inner.parts[HEAD].surface.commit();
-            self.buffers.push(buffer);
 
             // -> top-subsurface
-            let buffer =
-                pool.buffer(
-                    4 * (width * HEADER_SIZE) as i32,
-                    (width + 2 * BORDER_SIZE) as i32,
-                    BORDER_SIZE as i32,
-                    4 * (width + 2 * BORDER_SIZE) as i32,
-                    wl_shm::Format::Argb8888,
-                ).implement(|_, _| {});
+            let buffer = pool.buffer(
+                4 * (width * HEADER_SIZE) as i32,
+                (width + 2 * BORDER_SIZE) as i32,
+                BORDER_SIZE as i32,
+                4 * (width + 2 * BORDER_SIZE) as i32,
+                wl_shm::Format::Argb8888,
+            );
             self.inner.parts[TOP].subsurface.set_position(
                 -(BORDER_SIZE as i32),
                 -(HEADER_SIZE as i32 + BORDER_SIZE as i32),
@@ -533,17 +534,15 @@ impl Frame for BasicFrame {
                 );
             }
             self.inner.parts[TOP].surface.commit();
-            self.buffers.push(buffer);
 
             // -> bottom-subsurface
-            let buffer =
-                pool.buffer(
-                    4 * (width * HEADER_SIZE) as i32,
-                    (width + 2 * BORDER_SIZE) as i32,
-                    BORDER_SIZE as i32,
-                    4 * (width + 2 * BORDER_SIZE) as i32,
-                    wl_shm::Format::Argb8888,
-                ).implement(|_, _| {});
+            let buffer = pool.buffer(
+                4 * (width * HEADER_SIZE) as i32,
+                (width + 2 * BORDER_SIZE) as i32,
+                BORDER_SIZE as i32,
+                4 * (width + 2 * BORDER_SIZE) as i32,
+                wl_shm::Format::Argb8888,
+            );
             self.inner.parts[BOTTOM]
                 .subsurface
                 .set_position(-(BORDER_SIZE as i32), height as i32);
@@ -566,17 +565,15 @@ impl Frame for BasicFrame {
                 );
             }
             self.inner.parts[BOTTOM].surface.commit();
-            self.buffers.push(buffer);
 
             // -> left-subsurface
-            let buffer =
-                pool.buffer(
-                    4 * (width * HEADER_SIZE) as i32,
-                    BORDER_SIZE as i32,
-                    (height + HEADER_SIZE) as i32,
-                    4 * (BORDER_SIZE as i32),
-                    wl_shm::Format::Argb8888,
-                ).implement(|_, _| {});
+            let buffer = pool.buffer(
+                4 * (width * HEADER_SIZE) as i32,
+                BORDER_SIZE as i32,
+                (height + HEADER_SIZE) as i32,
+                4 * (BORDER_SIZE as i32),
+                wl_shm::Format::Argb8888,
+            );
             self.inner.parts[LEFT]
                 .subsurface
                 .set_position(-(BORDER_SIZE as i32), -(HEADER_SIZE as i32));
@@ -599,17 +596,15 @@ impl Frame for BasicFrame {
                 );
             }
             self.inner.parts[LEFT].surface.commit();
-            self.buffers.push(buffer);
 
             // -> right-subsurface
-            let buffer =
-                pool.buffer(
-                    4 * (width * HEADER_SIZE) as i32,
-                    BORDER_SIZE as i32,
-                    (height + HEADER_SIZE) as i32,
-                    4 * (BORDER_SIZE as i32),
-                    wl_shm::Format::Argb8888,
-                ).implement(|_, _| {});
+            let buffer = pool.buffer(
+                4 * (width * HEADER_SIZE) as i32,
+                BORDER_SIZE as i32,
+                (height + HEADER_SIZE) as i32,
+                4 * (BORDER_SIZE as i32),
+                wl_shm::Format::Argb8888,
+            );
             self.inner.parts[RIGHT]
                 .subsurface
                 .set_position(width as i32, -(HEADER_SIZE as i32));
@@ -632,10 +627,7 @@ impl Frame for BasicFrame {
                 );
             }
             self.inner.parts[RIGHT].surface.commit();
-            self.buffers.push(buffer);
         }
-        // swap the pool
-        self.pools.swap();
     }
 
     fn subtract_borders(&self, width: i32, height: i32) -> (i32, i32) {